requirements.txt
playwright==1.49.0
playwright-stealth==1.0.6
requests==2.32.3
python-dotenv==1.0.1
httpx==0.28.1
opencv-python==4.10.0.84
numpy==2.0.2
openpyxl==3.1.2

# Token Manager Dependencies
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
cryptography==43.0.1
websockets==13.1

# Testing Dependencies
pytest==8.3.3
hypothesis==6.112.0
